Feat  new dashboard exp (#24)
* adds league placeholder card

* cleans up league card details

* renames components

* renames files

* makes edits

* adds fix to show more page

* adds permissions in useEffect on render

* changes loading sinner

* cleans up code
