+++ src/components/dashboard/ActiveLeaguesSection.tsx
@@ -26,7 +26,7 @@ export const ActiveLeaguesSection: React.FC<ActiveLeaguesSectionProps> = ({
           <p>No active leagues available.</p>
         </div>
       ) : (
-        <div className="w-full grid grid-cols-1 lg:grid-cols-2 gap-3">
+        <div className="w-full grid grid-cols-1 gap-3 lg:gap-5">
           {activeLeagues.slice(0, 3).map((league, index) => (
             <LeagueCardDetailed
               key={index}
