+++ src/components/ui/LoadingState.tsx
@@ -1,3 +1,4 @@
+import { Loader } from "lucide-react";
 import { twMerge } from "tailwind-merge";
 
 interface LoadingStateProps {
@@ -7,7 +8,7 @@ interface LoadingStateProps {
 export const LoadingState = ({ message = "Loading..." }: LoadingStateProps) => {
   return (
     <div className="flex flex-col items-center justify-center py-12">
-      <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-primary-600 mb-4"></div>
+      <Loader className="animate-spin h-8 w-8 text-primary-500 mb-4"></Loader>
       <p className="text-gray-600 dark:text-gray-400">{message}</p>
     </div>
   );
